// src/interfaces/SysAPIParams.tsx
export interface InitialSysAPIParams {
    userEmail:     string;
    planId:        string;
    appId:         string;
    sysAPIUrl:     string;
    initialAPIKey: string;
  }

export interface SysAPIParams {
    userEmail:     string;
    planId:        string;
    appId:         string;
    sysAPIUrl:     string;
    sysAPIKey: string;
}